requirements.txt
Flask==2.2.2
Flask_Bootstrap==3.3.7.1
Flask_Login==0.6.2
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
Pillow==9.2.0
timeago==1.0.16
torch==1.12.0
torchvision==0.13.0
Werkzeug==2.2.2
WTForms==3.0.1
